-- add community controller and update community service

+++ app/src/main/java/org/closure/app/UserModule/models/UserModel.java
@@ -14,14 +14,14 @@ import lombok.With;
 @With
 public class UserModel {
 
-    private int id ;
+    private Long id ;
     private String name ;
     private String email ;
     private String password ;
     private String university ;
     private String img ;
     private String community_name ;
-    private int  study_year ;
+    private Integer  study_year ;
     private Date  start_year ;
     private Date  age ;
 }
